renamed ImageGrabberMixin to FitsHeaderMixin and derived class ImageFitsHeaderMixin for images
