+++ pyobs/mixins/__init__.py
@@ -9,3 +9,4 @@ from .follow import FollowMixin
 from .motionstatus import MotionStatusMixin
 from .waitformotion import WaitForMotionMixin
 from .weatheraware import WeatherAwareMixin
+from .fitsheader import FitsHeaderMixin
